requirements.txt
pyyaml==6.0.2
pydantic==2.8.2
confluent-kafka==2.5.3
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
pytest==8.3.2
